[1,2,3 단계] Web MVC - 김태정 (#19)
* [1단계 - 로그인] feat: 로그인 기능 - 토큰을 발급

* [1단계 - 로그인] feat: 내 정보 조회하기 - 토큰을 이용하여 본인 정보 응답하기

* [2단계 - 로그인 리팩터링] feat: 기본 데이터 저장 로직 작성

* [2단계 - 로그인 리팩터링] feat: 예약하기, 예약취소 개선 - 비로그인 사용자는 예약이 불가능하다.

* [2단계 - 로그인 리팩터링] feat: 예약하기, 예약취소 개선 - 자신의 예약이 아닌 경우 예약 취소가 불가능하다.

* [ 3단계 - 관리자 기능 보호] feat: 관리자 역할을 추가한다. 일반 멤버와 관리자 멤버를 구분한다.

* [3단계 - 관리자 기능 보호] feat: 관리자 역할을 추가한다. 일반 멤버와 관리자 멤버를 구분한다.

* [3단계 - 관리자 기능 보호] feat: 관리자 기능을 보호한다.
                        관리자 관련 기능 API는 /admin 붙이고 interceptor로 검증한다.
                        관리자 관련 기능 API는 authorization 헤더를 이용하여 인증과 인가를 진행한다.

+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -1,42 +1,58 @@
 package nextstep.reservation;
 
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
 import java.net.URI;
 import java.util.List;
+import java.util.Objects;
+import lombok.RequiredArgsConstructor;
+import nextstep.auth.AuthenticationException;
+import nextstep.auth.MemberAuthentication;
+import nextstep.member.Member;
+import nextstep.member.MemberService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequiredArgsConstructor
 @RequestMapping("/reservations")
 public class ReservationController {
 
-    public final ReservationService reservationService;
+  private final ReservationService reservationService;
+  private final MemberService memberService;
 
-    public ReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
+  @PostMapping
+  public ResponseEntity createReservation(@MemberAuthentication Member member,
+      @RequestBody ReservationRequest reservationRequest) {
+    Long id = reservationService.create(reservationRequest);
+    return ResponseEntity.created(URI.create("/reservations/" + id)).build();
+  }
+
+  @GetMapping
+  public ResponseEntity readReservations(@RequestParam Long themeId, @RequestParam String date) {
+    List<Reservation> results = reservationService.findAllByThemeIdAndDate(themeId, date);
+    return ResponseEntity.ok().body(results);
+  }
+
+  @DeleteMapping("/{id}")
+  public ResponseEntity deleteReservation(@MemberAuthentication Member member, @PathVariable Long id) {
+    var reservation = reservationService.findById(id);
+    if (!Objects.equals(reservation.getName(), member.getName())) {
+      throw new AuthenticationException("자신의 예약만 취소 가능합니다.");
     }
+    reservationService.deleteById(id);
 
-    @PostMapping
-    public ResponseEntity createReservation(@RequestBody ReservationRequest reservationRequest) {
-        Long id = reservationService.create(reservationRequest);
-        return ResponseEntity.created(URI.create("/reservations/" + id)).build();
-    }
+    return ResponseEntity.noContent().build();
+  }
 
-    @GetMapping
-    public ResponseEntity readReservations(@RequestParam Long themeId, @RequestParam String date) {
-        List<Reservation> results = reservationService.findAllByThemeIdAndDate(themeId, date);
-        return ResponseEntity.ok().body(results);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable Long id) {
-        reservationService.deleteById(id);
-
-        return ResponseEntity.noContent().build();
-    }
-
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity onException(Exception e) {
-        return ResponseEntity.badRequest().build();
-    }
+  @ExceptionHandler(RuntimeException.class)
+  public ResponseEntity<String> handle(RuntimeException exception) {
+    return ResponseEntity.badRequest().body(exception.getMessage());
+  }
 }
